# swagger (1).yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Online shop"
basePath: "/v1"
...
schemes:
- "http"
paths:
...
  /v1/Produk/{id_user}:
    get:
      tags:
      - "Produk"
      summary: " get produk berdasarkan id user"
      parameters:
      - name: "id_user"
        in: "path"
        description: "ID of pet to return"
        required: true
        type: "integer"
        format: "int64"
      description: ""
      responses:
        "200":
          description: "success"
    post:
      tags:
      - "Produk"
      summary: "Create produk"
      parameters:
      - name: "id_user"
        in: "path"
        description: "ID of pet to return"
        required: true
        type: "integer"
        format: "int64"
      description: ""
      responses:
        "200":
          description: "success"
    delete:
      tags:
      - "Produk"
      summary: "delete produk"
      parameters:
      - name: "id_user"
        in: "path"
        description: "ID of pet to return"
        required: true
        type: "integer"
        format: "int64"
      description: ""
      responses:
        "200":
          description: "success"
          
    put:
      tags:
      - "Produk"
      summary: "edit produk"
      parameters:
      - name: "id_user"
        in: "path"
        description: "ID of pet to return"
        required: true
        type: "integer"
        format: "int64"
      description: ""
      responses:
        "200":
          description: "success"
...
  /v1/penjualan/{id_user}:
    get:
      tags:
      - "Penjualan"
      summary: " get penjualan berdasarkan id user"
      parameters:
      - name: "id_user"
        in: "path"
        description: "ID of pet to return"
        required: true
        type: "integer"
        format: "int64"
      description: ""
      responses:
        "200":
          description: "success"
...